5/10 관리자 1:1 사진 제외 완, jquery 추가 및 수정 등...

+++ src/main/java/com/refactoring/rekall/service/UsQService.java
@@ -7,11 +7,13 @@ import com.refactoring.rekall.repository.UsQRepository;
 import lombok.RequiredArgsConstructor;
 import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -23,9 +25,12 @@ public class UsQService {
     @Autowired
     UserService userService;
 
-//  ---------------------------------- ★ 1:1 문의 저장★ ---------------------------------------------------------------
-    public void sendQuestion(UsQDTO usQDTO) {
-        usQRepository.save(UsQEntity.toUsQEntity(usQDTO));
+
+    //  ---------------------------------- ★ 1:1 문의 저장★ ---------------------------------------------------------------
+    public void saveQuestion(UsQDTO usQDTO) {
+        System.out.println("service : usqId "+usQDTO.getUsqId());
+        if(usQDTO != null)
+            usQRepository.save(UsQEntity.toUsQEntity(usQDTO));
     }
 
 //  ---------------------------------- ★ 1:1 문의 리스트★ ---------------------------------------------------------------
@@ -51,11 +56,53 @@ public class UsQService {
     }
 
 //  --------------------- ★ 1:1 문의 _ usQId로 삭제하기★ ---------------------------------------------------------------
-    public void deleteUsQ(List<Integer> usQIds) {
-        for(Integer usQ : usQIds) {
-            usQRepository.deleteById(usQ);
+    public void deleteUsQ(List<Integer> usQIds, Integer usqId, String page) {
+
+        if(page.equals("detail"))
+            usQRepository.deleteById(usqId);
+        else {
+            System.out.println("여기");
+            for (Integer usQ : usQIds) {
+                if(usQ != null) {
+                    System.out.println("usq" + usQ);
+                    usQRepository.deleteById(usQ);
+                }
+            }
         }
     }
 
+//  --------------------- ★ 전체 문의 내역★ ---------------------------------------------------------------
+    public List<UsQDTO> usqAllList() {
+        List<UsQEntity> usQEntityList = usQRepository.findAll(Sort.by(Sort.Direction.DESC, "usqId"));
+        List<UsQDTO> usQDTOList = new ArrayList<>();
+
+        for(UsQEntity usQEntity : usQEntityList) {
+            if(usQEntity != null) {
+                usQDTOList.add(UsQDTO.toUsQDTO(usQEntity));
+            }
+        }
+        return usQDTOList;
+    }
+//  --------------------- ★ 개인 문의 뽑기★ ---------------------------------------------------------------
+    public UsQDTO findUsQ(Integer usQId) {
+        Optional<UsQEntity> OpusQEntity = usQRepository.findById(usQId);
+        UsQDTO usQDTO = new UsQDTO();
+
+        if(OpusQEntity.isPresent()) {
+            usQDTO = UsQDTO.toUsQDTO(OpusQEntity.get());
+        }
+        return usQDTO;
+    }
+//  --------------------- ★ 카테고리별 뽑기★ ---------------------------------------------------------------
+    public List<UsQDTO> findCList(String categoryId) {
+        List<UsQEntity> UsQEntityList = usQRepository.findByCategoryEntityCategoryIdOrderByUsqIdDesc(categoryId);
+        List<UsQDTO> usQDTOList = new ArrayList<>();
+        for(UsQEntity usQEntity : UsQEntityList) {
+            if(usQEntity != null) {
+                usQDTOList.add(UsQDTO.toUsQDTO(usQEntity));
+            }
+        }
+        return usQDTOList;
+    }
 
 }
