+++ src/main/java/com/refactoring/rekall/service/WishListService.java
@@ -43,20 +43,45 @@ public class WishListService {
 //  ------------------------------------- ★ wishList 저장★ ------------------------------------------------------------
     public void getWishList(Integer productId, String loginId) {
         WishListDTO wishListDTO = new WishListDTO();
-        UserDTO userDTO = userService.findByUserID(loginId); // 회원정보 찾기
-        ProductDTO productDTO = productService.findByProductId(productId);
-        wishListDTO.setProductDTO(productDTO);
-        wishListDTO.setUserDTO(userDTO);
 
-        wishListRepository.save(WishListEntity.toWishListEntity(wishListDTO));
-        // 이상, 확인 필 저장은 되는데 화면이 X  확인해보기!
+        System.out.println("get wishList loginId" + loginId);
+        System.out.println("productId"+productId);
+        UserDTO userDTO = userService.findByUserID(loginId); // 회원정보 찾기
+        ProductDTO productDTO = productService.findByProductId(productId); // 상품
+
+        List<WishListDTO> wishListDTOList = findWishList(loginId); // 기존 회원의 wishList
+         if (!wishListDTOList.isEmpty()) {
+             boolean answer = true;
+             for(WishListDTO wishList : wishListDTOList) {
+                 if (wishList.getProductDTO().getProductId() != productDTO.getProductId()) {//기존 list에 없을 때 true
+                     answer = true;
+                     wishListDTO.setProductDTO(productDTO);
+                     wishListDTO.setUserDTO(userDTO);
+                 }   else { //  혹 중간에 있으면 false로 변환하고 for문 끝냄
+                     answer = false;
+                     break;
+                 }
+             }
+             if(answer)
+                 wishListRepository.save(WishListEntity.toWishListEntity(wishListDTO));
+         } else { // null이면 저장바로 하면 됨
+             System.out.println("null");
+             wishListDTO.setProductDTO(productDTO);
+             wishListDTO.setUserDTO(userDTO);
+             wishListRepository.save(WishListEntity.toWishListEntity(wishListDTO));
+         }
     }
 
 //  ------------------------------------- ★ wishList 삭제★ ------------------------------------------------------------
     public void delWishList(Integer productId, String userId) {
 
-      Object wishList  = wishListRepository.findWishListId(productId, userId);
-      Integer wishListId = (Integer)wishList;
-      wishListRepository.deleteById(wishListId);
+      System.out.println("delete" + productId+userId);
+      Integer wishListId   = wishListRepository.findWishListId(productId, userId);
+
+      System.out.println("wishListId"+wishListId);
+      if(wishListId != null)
+        wishListRepository.deleteById(wishListId);
+      else
+          System.out.println("null");
     }
 }
